fix: layout breaking on mobile devices
the reason for this is the use of flexbox to create the top level layout with sticky footer. Flexbox doesn't look white-space declarations a lot. The container ( a flex box) in this case must have a fixed width for the elements with white-space value other than normal to work properly.